Refactor all footer tests
